Remove mentions of the token on input
Also removes mentions of the token on output for now. This also  makes an assumption that the field provided as `origin` will be returned as `configUrl` on the IdentityCredential if the user selects an IdP configured for use as Lightweight by the RP. A separate field might be more appropriate if we think this field shouldn't be overloaded like this.